split out DecoderConfig
