+++ identity_jose/src/jws/mod.rs
@@ -9,6 +9,7 @@
 mod algorithm;
 mod charset;
 mod decoder;
+mod decoder_config;
 mod encoder;
 mod format;
 mod header;
@@ -18,6 +19,7 @@ mod recipient;
 pub use self::algorithm::*;
 pub use self::charset::*;
 pub use self::decoder::*;
+pub use self::decoder_config::*;
 pub use self::encoder::*;
 pub use self::format::*;
 pub use self::header::*;
