+++ identity_jose/src/tests/rfc8037.rs
@@ -6,8 +6,8 @@ use crypto::signatures::ed25519::SecretKey;
 use crate::jwk::Jwk;
 use crate::jws::Decoder;
 use crate::jws::Encoder;
-use crate::jws::JwsDecoderConfig;
 use crate::jws::JwsAlgorithm;
+use crate::jws::JwsDecoderConfig;
 use crate::jws::JwsHeader;
 use crate::jws::JwsSignatureVerifierFn;
 use crate::jws::Recipient;
