+++ identity_jose/src/tests/roundtrip.rs
@@ -10,8 +10,8 @@ use crate::jwk::JwkParamsOkp;
 use crate::jwk::JwkType;
 use crate::jws::Decoder;
 use crate::jws::Encoder;
-use crate::jws::JwsDecoderConfig;
 use crate::jws::JwsAlgorithm;
+use crate::jws::JwsDecoderConfig;
 use crate::jws::JwsHeader;
 use crate::jws::JwsSignatureVerifierFn;
 use crate::jws::Recipient;
